str_parameter.py
